feature : User, Schedule 권한 매핑

+++ src/main/java/com/example/schedule/user/service/UserService.java
@@ -7,6 +7,7 @@ import com.example.schedule.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import org.springframework.util.ObjectUtils;
 
 import java.util.List;
 
@@ -30,14 +31,21 @@ public class UserService {
     }
 
     @Transactional
-    public UserInfoResponse updateUserInfo(Long id, UpdateUserInfoRequest request) {
+    public UserInfoResponse updateUserInfo(Long id, UpdateUserInfoRequest request, Long loginId) {
 
         if (isUserExistsByEmailAndIdIsNot(request.getEmail(), id)) {
             throw new IllegalArgumentException("이미 사용중인 이메일입니다.");
         }
 
+        User loginUser = userRepository.findById(loginId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 회원입니다."));
+
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 회원입니다."));
+
+        if (!loginUser.isOwnerOf(user)) {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
 
         user.updateNameAndEmail(request.getName(), request.getEmail());
         userRepository.flush();
@@ -45,13 +53,22 @@ public class UserService {
     }
 
     @Transactional
-    public void deleteUserById(Long id) {
+    public void deleteUserById(Long id, Long loginId) {
+        User loginUser = userRepository.findById(loginId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 회원입니다."));
+
         User user = userRepository.findById(id)
                 .orElseThrow(() -> new IllegalArgumentException("존재하지 않은 회원입니다."));
+
+        if (!loginUser.isOwnerOf(user)) {
+            throw new IllegalArgumentException("권한이 없습니다.");
+        }
+
         userRepository.delete(user);
     }
 
     private boolean isUserExistsByEmailAndIdIsNot(String email, Long id) {
         return userRepository.existsByEmailAndIdIsNot(email, id);
     }
+
 }
